requirements.txt
Django==2.1.2
djangorestframework==3.8.2
django-rest-swagger==2.2.0
psycopg2-binary==2.7.5
gunicorn==19.9.0
whitenoise==4.1
pytest-django==3.4.3
pytest-xdist==1.23.2
graphene-django==2.2.0
pillow==5.3.0